refactor(replication): move object config bridge
