+++ crates/ecstore/src/bucket/replication/replication_object_bridge.rs
@@ -15,8 +15,8 @@
 use std::{collections::HashMap, sync::Arc};
 
 use super::replication_filemeta_boundary::{ReplicateDecision, ReplicationStatusType, ReplicationType};
+use super::replication_object_config::{check_replicate_delete, get_must_replicate_options, must_replicate};
 use super::replication_pool::{schedule_replication, schedule_replication_delete};
-use super::replication_resyncer::{check_replicate_delete, get_must_replicate_options, must_replicate};
 use super::replication_storage_boundary::{ObjectInfo, ObjectOptions, ObjectToDelete, ReplicationStorage};
 use rustfs_replication::{DeletedObjectReplicationInfo, MustReplicateOptions};
